# maglev: Parallelize calculation of permutations