@brb 2021-01-14
CI net-next hit https://github.com/cilium/cilium/issues/12511.